base the Makefile on the same template used for other packages
Beside the obvious differences (the package specific values for
(`PKG', `DEPS', `DOMAIN' and `MANUAL_HTML_ARGS) this only diverges
from the Makefiles of other packages that use the same template in
that an additional target `authors' is available and in that `EFLAGS'
is non-empty.

Compared to the old Makefile the most significant breaking change is
that there no longer exists a `bump-versions' target.  To bump the
versions one still only has to edit one file, but that's now the
library, not the Makefile.

Furthermore the autoloads file is now being created automatically, and
the manual can be exported to more formats and the results can easily
be published on the webpage.