with-editor-sleeping-editor: Also print the working directory
When `git-commit' is invoked from a sub-directory, then it changes the
working directory to the root of the working tree before invoking the
editor.  Additionally it ask the editor to edit a relative path in
most cases.  That path is relative to the top-level, so we need the
sleeping editor to also print the name of that directory; or we would
end up trying to edit e.g. "/path/to/repo/subdir/.git/COMMIT_EDITMSG"
instead of "/path/to/repo/.git/COMMIT_EDITMSG".

Actually, as of 2.19.0, `git-commit' no longer does this, it uses an
absolute file-name now.  But programs might still do it and that is
legitimate.  `$EDITOR' inherits the working directory from its parent
process, so it normally isn't a problem if the file-name is relative.
It only a problem here because `$EDITOR' in hour case forwards the
edit request to a running `emacs' instance, which is not a child
process and therefore does not inherit the working directory.

Fixes #55.